Migrate more FASTA loaders, most of ProteinManager, and ProteinController
