+++ ms2/src/org/labkey/ms2/AnnotationView.java
@@ -17,10 +17,10 @@ package org.labkey.ms2;
 
 import org.apache.commons.collections4.MultiValuedMap;
 import org.labkey.api.protein.IdentifierType;
+import org.labkey.api.protein.ProteinManager;
 import org.labkey.api.util.HtmlString;
 import org.labkey.api.view.JspView;
 import org.labkey.ms2.protein.Protein;
-import org.labkey.ms2.protein.ProteinManager;
 
 import java.util.Collection;
 import java.util.Collections;
