+++ ms2/src/org/labkey/ms2/MS2Module.java
@@ -32,10 +32,13 @@ import org.labkey.api.module.SpringModule;
 import org.labkey.api.ms2.MS2Service;
 import org.labkey.api.pipeline.PipelineService;
 import org.labkey.api.protein.CustomAnnotationSetManager;
+import org.labkey.api.protein.CustomProteinListView;
 import org.labkey.api.protein.ProteinAnnotationPipelineProvider;
+import org.labkey.api.protein.ProteinController;
 import org.labkey.api.protein.ProteinSchema;
 import org.labkey.api.protein.ProteinService;
 import org.labkey.api.protein.ProteomicsModule;
+import org.labkey.api.protein.fasta.FastaDbLoader;
 import org.labkey.api.protein.query.CustomAnnotationSchema;
 import org.labkey.api.protein.query.ProteinUserSchema;
 import org.labkey.api.protein.query.SequencesTableInfo;
@@ -76,10 +79,7 @@ import org.labkey.ms2.pipeline.sequest.SequestPipelineProvider;
 import org.labkey.ms2.pipeline.sequest.SequestSearchTask;
 import org.labkey.ms2.pipeline.sequest.ThermoSequestParamsBuilder;
 import org.labkey.ms2.pipeline.tandem.XTandemPipelineProvider;
-import org.labkey.ms2.protein.CustomProteinListView;
-import org.labkey.ms2.protein.FastaDbLoader;
 import org.labkey.ms2.protein.Protein;
-import org.labkey.ms2.protein.ProteinController;
 import org.labkey.ms2.protein.ProteinServiceImpl;
 import org.labkey.ms2.query.MS2Schema;
 import org.labkey.ms2.reader.DatDocumentParser;
