+++ ms2/src/org/labkey/ms2/PeptideImporter.java
@@ -23,10 +23,10 @@ import org.labkey.api.data.SqlSelector;
 import org.labkey.api.data.Table;
 import org.labkey.api.exp.XarContext;
 import org.labkey.api.protein.ProteinSchema;
+import org.labkey.api.protein.fasta.FastaDbLoader;
 import org.labkey.api.query.FieldKey;
 import org.labkey.api.security.User;
 import org.labkey.api.util.PageFlowUtil;
-import org.labkey.ms2.protein.FastaDbLoader;
 import org.labkey.ms2.reader.AbstractQuantAnalysisResult;
 import org.labkey.ms2.reader.MS2Loader;
 import org.labkey.ms2.reader.PeptideProphetHandler;
