+++ ms2/src/org/labkey/ms2/protein/uploadCustomProteinAnnotations.jsp
@@ -15,12 +15,12 @@
  * limitations under the License.
  */
 %>
+<%@ page import="org.labkey.api.protein.CustomAnnotationType" %>
+<%@ page import="org.labkey.api.protein.ProteinController.UploadAnnotationsForm" %>
+<%@ page import="org.labkey.api.protein.ProteinController.UploadCustomProteinAnnotations" %>
 <%@ page import="org.labkey.api.view.HttpView" %>
 <%@ page import="org.labkey.api.view.JspView" %>
 <%@ page import="org.labkey.api.view.template.ClientDependencies" %>
-<%@ page import="org.labkey.api.protein.CustomAnnotationType" %>
-<%@ page import="org.labkey.ms2.protein.ProteinController.UploadAnnotationsForm" %>
-<%@ page import="org.labkey.ms2.protein.ProteinController.UploadCustomProteinAnnotations" %>
 <%@ page extends="org.labkey.api.jsp.JspBase"%>
 <%@ taglib prefix="labkey" uri="http://www.labkey.org/taglib" %>
 <%!
