+++ ms2/src/org/labkey/ms2/runSummary.jsp
@@ -17,13 +17,13 @@
 %>
 <%@ page import="org.apache.commons.lang3.StringUtils" %>
 <%@ page import="org.labkey.api.data.Container" %>
+<%@ page import="org.labkey.api.protein.ProteinManager" %>
 <%@ page import="org.labkey.api.protein.fasta.FastaFile" %>
 <%@ page import="org.labkey.api.view.HttpView" %>
 <%@ page import="org.labkey.api.view.JspView" %>
 <%@ page import="org.labkey.ms2.MS2Controller" %>
 <%@ page import="org.labkey.ms2.MS2Run" %>
 <%@ page import="org.labkey.ms2.pipeline.mascot.MascotRun" %>
-<%@ page import="org.labkey.ms2.protein.ProteinManager" %>
 <%@ page import="java.util.ArrayList" %>
 <%@ page import="java.util.List" %>
 <%@ page extends="org.labkey.api.jsp.JspBase" %>
